fix(build): resolve redefinition error and format string mismatch
- Removed inline getName() definition from NeuralEmbeddingNode header.
- Explicitly cast duration count to long long in LOGI to fix printf format warning.
- Ensures host tests compile correctly in GitHub Actions Run #585.

+++ src/capabilities/neural_embedding_node.cpp
@@ -146,7 +146,7 @@ std::vector<float> NeuralEmbeddingNode::generateEmbedding(const std::string& inp
         return {};
     }
     auto end = std::chrono::high_resolution_clock::now();
-    LOGI(TAG, "Inference Latency: %lld ms", std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count());
+    LOGI(TAG, "Inference Latency: %lld ms", (long long)std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count());
 
     // Extract embedding (assuming average pooling or [CLS] token at index 0)
     int output_idx = m_impl->interpreter->outputs()[0];
